@@ -4,14 +4,14 @@ apply plugin: 'com.github.dcendents.android-maven'
 group='com.github.TutorialsAndroid'
 
 android {
-    compileSdkVersion 28
+    compileSdkVersion 29
     resourcePrefix 'crash_'
 
     defaultConfig {
         minSdkVersion 15
-        targetSdkVersion 28
-        versionCode 4
-        versionName "1.0.19" //renamed library version properly.
+        targetSdkVersion 29
+        versionCode 5
+        versionName "2.0.19" //renamed library version properly.
     }
 	buildTypes {
         release {
@@ -25,10 +25,10 @@ dependencies {
 	implementation fileTree(include: ['*.jar'], dir: 'libs')
 
     //ANDROID SUPPORT LIBRARIES
-    implementation 'com.android.support:appcompat-v7:28.0.0'
-    implementation 'com.android.support:support-v4:28.0.0'
-    implementation 'com.android.support:recyclerview-v7:28.0.0'
-    implementation 'com.android.support:design:28.0.0'
-    implementation "com.android.support:preference-v7:28.0.0"
-    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
+    implementation 'androidx.appcompat:appcompat:1.0.0'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'androidx.recyclerview:recyclerview:1.0.0'
+    implementation 'com.google.android.material:material:1.0.0'
+    implementation 'androidx.preference:preference:1.0.0'
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
 }
